chore(ux): ai first updates (#50192)
