+++ frontend/src/layout/panel-layout/ai-first/NavLink.tsx
@@ -21,7 +21,7 @@ export function NavLink({ to, label, icon, isCollapsed }: NavLinkProps): JSX.Ele
     const { showConfigurePinnedTabsModal } = useActions(navigationLogic)
     const { pathname } = useValues(panelLayoutLogic)
 
-    const isHomePage = to === urls.projectHomepage()
+    const isHomePage = to === urls.projectRoot()
     const currentPath = removeProjectIdIfPresent(pathname)
     const isActive = currentPath === to
