+++ frontend/src/layout/panel-layout/panelLayoutLogic.tsx
@@ -15,6 +15,7 @@ export type PanelLayoutNavIdentifier =
     | 'Shortcuts'
     | 'DataManagement'
     | 'DataAndPeople'
+export type NavExperimentTab = 'home' | 'chat'
 export type PanelLayoutTreeRef = React.RefObject<LemonTreeRef> | null
 export type PanelLayoutMainContentRef = React.RefObject<HTMLElement> | null
 export const PANEL_LAYOUT_DEFAULT_WIDTH: number = 245
@@ -44,6 +45,7 @@ export const panelLayoutLogic = kea<panelLayoutLogicType>([
         setMainContentRect: (rect: DOMRect) => ({ rect }),
         setSidePanelWidth: (width: number) => ({ width }),
         toggleNavSection: (section: string) => ({ section }),
+        setNavExperimentTab: (tab: NavExperimentTab) => ({ tab }),
     }),
     reducers({
         isLayoutNavbarVisibleForDesktop: [
@@ -144,6 +146,13 @@ export const panelLayoutLogic = kea<panelLayoutLogicType>([
             0,
             {
                 setSidePanelWidth: (_, { width }) => width,
+            },
+        ],
+        navExperimentActiveTab: [
+            'home' as NavExperimentTab,
+            { persist: true },
+            {
+                setNavExperimentTab: (_, { tab }) => tab,
             },
         ],
         expandedNavSections: [
