+++ frontend/src/scenes/max/components/AiFirstMaxInstance.tsx
@@ -3,6 +3,7 @@ import { BindLogic, useActions, useValues } from 'kea'
 import { IconOpenSidebar, IconShare } from '@posthog/icons'
 import { LemonBanner } from '@posthog/lemon-ui'
 
+import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { copyToClipboard } from 'lib/utils/copyToClipboard'
 import { sceneLogic } from 'scenes/sceneLogic'
@@ -78,6 +79,7 @@ interface AiFirstMaxInstanceProps {
 export function AiFirstMaxInstance({ tabId }: AiFirstMaxInstanceProps): JSX.Element {
     const { threadVisible, threadLogicKey, conversation, conversationId } = useValues(maxLogic({ tabId }))
     const { startNewConversation } = useActions(maxLogic({ tabId }))
+    const isAIFirst = useFeatureFlag('AI_FIRST')
 
     const threadProps: MaxThreadLogicProps = {
         tabId,
@@ -87,7 +89,7 @@ export function AiFirstMaxInstance({ tabId }: AiFirstMaxInstanceProps): JSX.Elem
 
     return (
         <div className="flex grow overflow-hidden h-full">
-            <ChatHistoryPanel tabId={tabId} />
+            {!isAIFirst && <ChatHistoryPanel tabId={tabId} />}
             <BindLogic logic={maxLogic} props={{ tabId }}>
                 <BindLogic logic={maxThreadLogic} props={threadProps}>
                     <div className="flex flex-col grow overflow-hidden">
